Bug#36292498: Compile MySQL with clang 18 [noclose]
Building with libc++ 18 gives some deprecation warnings for usages of
std::basic_string<unsigned char>:

warning: 'char_traits<unsigned char>' is deprecated: char_traits<T>
for T not equal to char, wchar_t, char8_t, char16_t or char32_t is
non-standard and is provided for a temporary period. It will be
removed in LLVM 19, so please migrate off of it.

The usages that cause these warnings are unused type aliases. Since
they are unused, they are simply removed to silence the warnings.

Change-Id: I77bd84c002322d16e69e7b94babbd7cea6311c83