Bug#36393211 Error due to router-id limit
The router does not allow the router_id higher than 999999 in order
to be able to create an user name containing mysql_routerXX_YY where
XX is the router_id and YY is random 12-character suffix.

This patch changes that so that the highest possible router_id is
supported (2^32-1). To allow that the YY part is reduced to
7 characters.

Change-Id: I7db486d8849b01a790e07f265941c9fd89d27309