Bug#36399256 Excessive RAM usage causing disk swapping building mysql on windows with Ninja
If you use Ninja on windows (as VS.code does) it will try to
parallelize as much of the build process as it can.  This includes
linking binaries.

And, on Windows, this can take *A LOT* of memory.

The huge binaries are 'mysqld' and all the "large" unit tests,
i.e. the ones which link with the entire server in the form of
server_unittest_library.

The solution is to create a JOB_POOL of size one for linking these
executables.

Change-Id: Id9977db0b7a17b0f7c6a27196cbd91f9717ec02e