Bug#36275182: Thread pool deadlock seen when setting server offline mode
Problem: SET offline_mode=on could hang and make system unresponsive
Root cause: With many connections SET offline_mode could end up writing
so much to the TP notify descriptor that send() would block. If this
happens when the TP waiting_thread is waiting for a mutex held by SET
offline_mode a deadlock would result.

Solution: Add a flag which is set when the notify descriptor is written
to, and cleared by the waiting thread when it drains (reads all
available bytes) the notify descriptor. As long as the flag is set new
kill notifcations do not actually write to the notify descriptor. That
way the capacity of the notify descriptor is not exceeded.
Additional writes before the notify descriptor has been drained are
redundant since the waiting_thread iterates over
all connections in the thread group to identify those that are killed.

Increase the number of connections allowed by mysqltest, to be able to
create the required number of connections.

Change-Id: I8c84c0b9c9e5a4e2312fd6758e07eebc18e7c7a4