Bug#36356279: possible null pointer dereference in Item_ref::real_item() Bug#36204344: Assertion `m_ref_count > 0' failed|sql/item.h
When incrementing reference count for an expression, underlying expressions
within this expression are not looked at. However, while removing an
expression, after decrementing the reference count, even the underlying
expressions are looked at. This leads to deletion of the underlying
expression when they are not supposed be.
Solution is to not look at the underlying expression unless the current
expression is the last reference left.

Change-Id: I404c17b69ecd4e0e355b0650d6d3b654d003783c