Bug#33390949: Execute a statement with secondary engine, two entries in performance schema [2/2, prepared statements]
Prepared statements did not record in the
performance_schema.events_statements_* tables whether they had been
executed in a secondary engine. They did record it in
performance_schema.prepared_statements_instances, but not always. If
the statement went through the multi-engine optimization, it was
recorded correctly, but if it did not (for example because
use_secondary_engine was FORCED or OFF, or because the estimated cost
of executing in the primary engine was below
secondary_engine_cost_threshold), it was recorded as if it was
executed in the same engine as in the previous execution.

Fixed by moving the code that updates secondary engine information in
performance schema to a different place in
Prepared_statement::execute_loop(), so that it is always set to the
engine actually used in the execution. Also added calls to update the
events_statements_* tables there.

Change-Id: I3bf50f1c0821c7d27275cdd9335463f0883487b6