Bug#36423593 Purging ndb_binlog_index affects ndb binlog
Problem:
When a binary log file is purged and removed from disk there is also a
need to remove the rows corresponding to that file from the
mysql.ndb_binlog_index system table. When there are many rows to delete
such removal might take several seconds, thus affecting ndb binlog
handling of data changes from the NDB cluster.

Analysis:
The binlog MySQL Server can be seen to slow down and appear stalled at
the same time as purge is done. Purging the binary log files normally
have a small effect on the binlog rate, but should not be of this
degree.

The ndb_binlog_index table is stored in the InnoDB engine which can
be configured to achieve slight improvement. When slow down occurs
warning messages regarding redo log capacity can be seen, this indicates
that also slow disks or low number of available cpu cycles
should be considered when solving this problem.

Deletion of rows are done by deleting all rows where the 'File' column
matches the name of the removed file. By default there are no index on
the 'File' column and thus this results in a full table scan. Some
improvements can be achieved by creating an index on the column:

  mysql> CREATE INDEX ix_file ON mysql.ndb_binlog_index (File);

But even with an index the ndb binlog capacity is affected and thus the
rows should be removed using a different strategy.

Solution:
To avoid waiting for rows to be deleted all at once they are now
asynchronously removed piece by piece in the background. This is done
by adding a new Purger component with a dedicated thread that allows
offloading the removal of the rows.

Purges done using PURGE BINARY LOGS will also use the Purger but still
act synchronously and not return until all rows have been removed. By
using the Purger the NDB binlog injector capacity is not affected.

Rows potentially referencing already purged binary log files should not
be kept in ndb_binlog_index. In order to avoid such orphan rows,
consistency is checked on MySQL Server startup and
rows referencing non existing binary log files are removed.

The rate of removing rows can be controlled using the new system
variable --ndb-log-purge-rate= which allows controlling
the number of rows deleted per batch. The default value is set to 8192
and should work sufficiently. There is a delay of 100ms between each
batch. It's not possible to disable the asynch
purging.

Three new status variables "Ndb_log_purged_rows",
"Ndb_log_purged_files" and "Ndb_log_purge_queue_size" are added to allow
monitoring the number of rows and files purged by the Purger.

This patch also fixes Bug#36420628 by using committed read in order to
avoid locking the whole table against inserts while looking for rows to
delete.

Change-Id: I514aa530be7a0819634bd737e758ad9028d763d4