Bug#36367313: String::length at include/sql_string.h Bug#36367776: Assertion failure in Item_func_numhybrid::val_str
These bugs are two different problems with a consolidated solution.

Bug#36367313 hits an item that is not yet fixed when searching for
an item placed in the GROUP BY list. The solution to this is to stop
performing a full string evaluation in
Item_func_date_format::eq_specific(), but rather do a more selective
check for Item_string classes.

Bug#36367776 is also related to the TIME_FORMAT function, in this case
it complains about an argument that is NULL, and there is missing
check for nullptr after evaluation.

The solution is in three parts:

First, the sequence of checks in Item_func::eq() is changed so that we
always call eq_specific() last. That way, we have always performed a
reasonably thorough check of the function's arguments first and can
simplify the solution in eq_specific().

Second, Item_func_date_format::eq_specific() is modified so that we
only perform an extra check when the second argument is a literal
string value. Notice also the "return false" that is changed to
"return true", since the other cases have already been checked.

Third, implement Item_string::eq_binary() that compares the strings
in two Item_string object byte for byte.

Change-Id: Ib6a16499ddf40ce4a93e625a7c3d22fdce5b05b6