Bug#36446604 Metadata component does not log with thread ID [thread_id]
Step 3

Problem:

When the Metadata component of the ndbcluster plugin outputs a log
line, its THD id should be != 0, to indicate that a thread has a
correctly assigned THD handle. This number helps trace through the
perf_schema.error_log and the ndbinfo views (under
mysql_connection_id).

Analysis:

On setup of each THD, the thread id should be assigned from the global
atomic counter of the Global_THD_manager. If it doesn't it uses the
reserved_thread_id (0).

Solution:

The Ndb_thd_guard setup of THD now has the `set_new_thread_id()` and
the "THD::m_command" member variable is set to COM_DEAMON (as it
would otherwise have been uninitialized).

The Ndb_metadata component will now log a message when started
and every time it has connected to NDB.

Change-Id: I292ce85bfa01e6b3e15c3793a4af4cd5a3c816a7