Bug#36421152 the $expect_crash_on_start mechanism is racy version of restart_abort
This patch replaces usages of
```
--let $restart_parameters = restart: $params
--let $expect_crash_on_start = 1
--source include/start_mysqld.inc
```
pattern with a more robust
```
--exec echo "restart_abort: $params" > $_expect_file_name
```
supported natively by mysql-test-run.pl.
To help guard against common pitfalls a wrapper script with sanity
checks can be used instead:
```
--let $restart_parameters = $params
--source include/start_mysqld_expecting_crash.inc
```

This patch also fixes a wrong usages of Perl's sleep with a subsecond argument.
The sleep EXPR treats EXPR as integer, so sleep 0.1 or sleep 0.5 become sleep 0.
This patch replaces that with sleep 1, adjusting the number of iterations.
An alternative would be to use usleep, but this would require Time::HiRes.

Change-Id: I2c53a47a574948d90dd2420988da80b66021d8eb