Bug#36394602 - Plugin mysqlx reported: 'Startup failed with error "Server API not ready"'
Description
===========
There is an race condition between MySQL SQL Service API,
and X Plugin startup code.
When the server shutdown is fast enough it make mark
SQL Service is not avaiable, event if X Plugin gets
the information that the service is ready.

Fix
===
Revert the implementation of Bug#36245953, and reimplement it
using safer mechanisms.

Change-Id: I25601c5ace4abe1207d5e31ecc8fe41ebba7dd54