Bug#33057164 PSI_THREAD_INFO not updated while executing PREPARED STATEMENTS Bug 104121 PSI_THREAD_INFO not updated while executing PREPARED STATEMENTS
Problem:

When executing a PREPARED statement,
table performance_schema.threads does not populate the PROCESSLIST_INFO
column.

This prevents to properly observe the execution of prepared statements.

This is also a regression, introduced in:
- MySQL 5.6.46
- MySQL 5.7.28
- MySQL 8.0.18

Root cause:

The fix for:
  Bug 20712046 SHOW PROCESSLIST AND PERFORMANCE_SCHEMA TABLES DO NOT MASK PASSWORD FROM QUERY
removed a call to PSI_THREAD_CALL(set_thread_info) located in THD::set_query()

PSI_THREAD_CALL(set_thread_info) was called from different places in the
same fix, to make sure a properly sanitized query text is displayed if
needed, but the code path for prepared statements was overlooked.

Fix:

The fix is to call PSI_THREAD_CALL(set_thread_info) in the missing code
path, when executing a prepared statement.

The call uses either the executed query, or a sanitized version of it
if it contains sensitive data.

Additional fix:

The logic around rewrite_query_if_needed() was revised.

Trying to anticipate if a query will be displayed or not,
to possibly save a rewrite, is too fragile, and too dangerous.

Queries that contain sensitive data are always rewritten,
regardless of whether the query will actually be printed somewhere or not.

Change-Id: Ief9c80a24f539f0889371766843d2f7b05812ad8