Bug#36343647 [ERROR] [MY-013183] [InnoDB] Assertion failure:ibuf0ibuf.cc:3833:ib::fatal
While investigating above bug, we noticed the test was also failing
because of the following bug which is fixed in 8.4. Therefore,
cherry-picked following commit#79d678d8219 from 8.4 to 8.0 :

Bug#34982348 Assertion failure: mtr0mtr.cc:310:ib::fatal

Background:
-----------
At the time of inserting a record in the secondary index, innodb
tries to find out the position on the tree where to insert the record.
During this time if innodb detects that required page is not in the
buffer pool then it tries to insert the record in the change buffer
(ibuf) while holding an S latch on the tree. On the other hand, once
change buffer has buffered that entry in its ibuf tree, it detects if
the ibuf index tree needs to be modified then it tries to contract
the ibuf tree. In the process, it calls the ibuf_merge_pages() method
where we added the `log_free_check()` method before starting an mtr.

`log_free_check()` method was added through WL#10310. It has novel
intent that is it ensures that there is sufficient space in the redo
log buffer and, (debug only) thread is not holding any latch.

In this case query thread is holding the S-latch on the index tree.
`log_free_check()` must not be called with active latches as that may
lead to deadlocks.

But merges happen through the background master thread as well. At this
time the thread is not already inside a parent mtr.
Does that mean we need the log_free_check() in this case ?

Notice :
* ibuf_merge_in_background() calls ibuf_merge_pages() with sync=false
* mtr created in the ibuf_merge_pages() is read-only
* With the sync=true the buf_read_ibuf_merge_pages() call does
  not exit until it reads secondary index leaf pages mentioned in the
  ibuf pages. These reads may in turn cause the ibuf merge. Now, these
  ibuf merges may generate more redo logs irrespective of current or
  io thread.

That means the following :

- if the current thread is background thread then it doesn't require
  `log_free_check()` because it does the merges with sync=false.

- if the current thread is a query thread that is always inside an mtr,
  then in this case too it doesn't require `log_free_check()`.
  This thread may do the merges synchronously as well.

Fix:
----
- Removed the ` log_free_check()` call from the `ibuf_merge_pages()`
- Introduced a wrapper method that help us detect situation if the
  thread is already inside an mtr.
- Added relevant asserts to validate observations mentioned before.
- Developed an mtr test to verify ibuf contraction.

Change-Id: I44a1bf7e4605e51e485576807c8578f9c769d993