Bug#35854919 Client sessions not killed during shutdown, server seems stuck
During shutdown, killing connections or forcefully disconnecting them
is done based on the THDs. However, running with the threadpool plugin
may leave connections in a queue not yet handled by the thread group's
conncection handling thread, when the server starts shutting down.
Thus, the killing and forceful disconnect of the connections may be
completed before the THD for the queued connection request is created,
and hence, when the THD is eventually constructed, it will neither be
killed nor forecfully disconnected, but still, the server will wait
for the global connection count to be decremented, because the count
was incremented when the request was received by the connection event
loop. However, the counter will of course never be decremented as long
as the client keeps the connection open.

The suggested fix is to let the connection thread in each thread group
just abandon unhandled connection requests when the server has aborted
its connection event loop.

Change-Id: I7c8d2ea4407a23d4486cbd46040846a576d67d81