Bug#35507223 MySQL Server crashes when executing query
ISSUE
=====
The expression ~NO_ACCESS was being used to provide an "all access"
ACL bitmask.  This usage was successful when an ACL bitmask was a 32
bit value, but can cause problems now an ACL bitmask is a 64 bit value
(on Linux, not on Windows) as ~NO_ACCESS produces a value with the
high order 32 bits set.

SOLUTION
========
Make the type of all ACL bitmask variables explicitly 32 bits, by
changing the relevant bitmask variable types from ulong to uint32_t.

Change-Id: I7abe9ac400490de1d855a23ff5ced1c9124f1453