Bug#36440609 Reinstate Group Replication Tests with PAXOS Single Leader enabled
Description:
-----------
Group Replication test suite has some tests failing when we enabled
PAXOS Single Leader.

In order to solve and diagnose issues regarding its usage, we need to have a test suite
running regularly with PAXOS Single Leader = ON

Analysis and Fix:
----------

All failures are related with the tests themselves, due to not preserving
the paxos_single_leader variable value between restarts or not accounting
for it to be enabled.

Change-Id: I106f0ea5d37f6ec0b6b06a6321a00fe2671ebf5b