Bug#36452427 add Windows versioninfo resource to Router binaries
On windows all shared libraries and executables are expected to contain

- version-info
- product-name
- copyright notice

For the server this is implemented via a resource-file which is added
via ADD_VERSION_INFO() in cmake.

As Router is its own product, it needs its own resource-file to be
linked into all shared libraries, plugins and executables.

Change
======

- added extra argument to ADD_VERSION_INFO() select resource-files
  per cmake-COMPONENT
- added ROUTER_ADD_SHARED_LIBRARY() as wrapper around
  ADD_SHARED_LIBRARY()
- added NAMELINK_SKIP to ADD_SHARED_LIBRARY()
- refactored all router's shared libraries to use
  MYSQL_ADD_SHARED_LIBRARY() to simplify creating libraries and add
  version-info to all shared libs
- renamed harness-library.ddl to mysqlharness.dll

Change-Id: I8d700c22ec1522d9aaae80685eed3de3d490a12d