Bug#36426461 For users with OpenSSL 1.0.x, NDB TLS should fail
Until now, the config parameter RequireTls has been ignored when
NDB is compiled with OpenSSL 1.0.x, to allow certain tests to
pass on older platforms. With this change, RequireTls will not
be ignored, so tests will fail. A related change in bug#36493264
will disable TLS-related MTR tests on the unsupported platforms.

Change-Id: Idb1b01adf3686430715879f5e879b48e5c2a6827