Bug#36443995 Mysqld crash-Sig 11- mem_root_deque<QueryLevel>::clear with hypergraph optimizer
Also closes duplicate Bug#36107656.

Problem: this is a regression from

commit 6c388a7b4e1b29ec2a8d33aaa7a8ddadb6675dc8
Author: Gleb Shchepa <gleb.shchepa@oracle.com>
Date:   Wed Jan 9 21:34:27 2019 +0400

    Bug #28968848: SIMPLIFY MYSQL_YYABORT

which introduced the code in THD::cleanup_after_parse_error. This
destroys LEX::sp_head which destroys the parser's MEM_ROOT. When we
have a stored procedure with a recursive/nested parse, which we do in
this case, caused by the CTE which leads to the call of
reparse_common_table_expr, and we have an error in the CTE (in this
case LIMIT & IN/ALL/ANY/SOME in subquery), we call
cleanup_after_parse_error in nested parser, which destroys sp_head,
which in turn destroys the MEM_ROOT used *also for the top level
parse*. On the way out the error code path, we therefore find
ourselves in the code of objects that have been destroyed and thrashed
and also accessing their member variables. This leads to the crash.

It happens also without hypergraph optimizer, cf. Chaithra's simplified
repro.

Solution: make sure we only call sp_head::destroy(sp) in the top level
parse, so we don't saw off the branch we are sitting on.

Change-Id: I9ae75b9e94ac12940ca4e83f3fa82392a4c01108