Bug#36475850 ndb_rpl_log_updates and ndb_rpl_log_updates_filter fails in PB2
Test does not wait for data changes to be replicated before performing
validations. The only wait is for own server to binlog the changes and
that does not mean it has been replicated to the other clusters.

Fix by waiting for DDL and DML changes to propagate through the
replication topology before  performing validations. Remove wait for own
server to binlog.

Change-Id: Iad4ed872453bfa1cfcdcbbaf5127e77f47a249a0