BUG#36485520 Binlog index purge callback only for binlog
Problem:
The handlerton callback `ha_binlog_index_purge_file()` is called both
when purging binary log and relay log files. This is not correct as the
callback is only intended to be called when removing a binary log file
and the binlog index is updated, not when removing relay log files.

Analysis:
The purpose of the callback is to allow the ndbcluster plugin to remove
its record of binary log file coordinates which are kept in the
mysql.ndb_binlog_index table. The ndbcluster plugin keep no record
related to relay log files and thus there is nothing to do when a relay
log file is purged.
The callback is not used by any other storage engine and the callback
interface is not considered public. Also the naming implies that the
callback should only relate to "binlog and the index".

Fix:
Call the ha_ndbcluster_binlog_index() callback only as part of purging a
binary log file.

Change-Id: I32efffa90388360cbbbe8360a3d4407cd77ee186