Bug#36189820 Server crash happened while running 'explain format=json for           connection xxx [follow-up]
Undo a too aggressive cleanup done as part of the original patch:
would fail on a query in subquery_bugs in which case, the iterator was
initialized with an empty root.

We had replaced a check of null m_current with an assert in
Query_term_iterator &operator++(), which we hit with the mentioned
query. Undo that and bring back the null test.

Change-Id: I4f1ea0de0b515bd318ee6a3dcf225378a1bdf261