Bug#36495702: Speed up main.hypergraph_bugs
One of the queries in the test case for Bug#36099491 took a long time
to run in debug builds. That single query took around eight times as
much time to run as all the other queries in the test together.

Fixed by only explaining the query instead of executing it, as EXPLAIN
is sufficient to reproduce the original bug.

Change-Id: I07d0810933ebc59b743465c64a3d5d97f19a5e7a