Bug#36496160 Incorrect results when using distinct and order by and derived table
When de-duplicating the rows, only the first order by expression, a
column, is used.  The function call (duplicated in the select list) is
not added to the set of ORDER expressions used for computing the
group, so only two rows ever make it to the actual sorting as seen in
the following EXPLAIN ANALYZE:

-> Sort: t0.`name`, `str_to_date(tt.``time``,'%Y%m%d')`  (actual time=0.876..0.877 rows=2 loops=1)
    -> Table scan on <temporary>  (cost=2.02..4.24 rows=7) (actual time=0.769..0.774 rows=2 loops=1)
        -> Temporary table with deduplication  (cost=1.65..1.65 rows=7) (actual time=0.76..0.76 rows=2 loops=1)
            -> Table scan on tt  (cost=0.95 rows=7) (actual time=0.42..0.539 rows=7 loops=1)

If we use an alias for the equivalent call to in the select list, the error is not seen.

Checking the optimizer trace, we see this here:

              "simplifying_group_by": {
                "original_clause": "`tt`.`name`",
                "items": [
                  {
                    "item": "`tt`.`name`"
                  }
                ],
                "resulting_clause_is_simple": true,
                "resulting_clause": "`tt`.`name`"
              }

The function call is missing, whereas if we use an alias instead for
the function call, we see:

              "simplifying_group_by": {
                "original_clause": "`tt`.`name`,`b`",
                "items": [
                  {
                    "item": "`tt`.`name`"
                  },
                  {
                    "item": "str_to_date(`tt`.`time`,'%Y%m%d')"
                  }
                ],
                "resulting_clause_is_simple": true,
                "resulting_clause": "`tt`.`name`,`b`"
              },

Checking the code in create_order_from_distinct, we
see that the criterion for adding an expression to the group by list is

   order->in_field_list

But the expression in the ORDER BY clause has been added as a hidden
item to Query_block::fields, not being recognized as being the same as
the equivalent function call in the select list, so

   order->in_field_list == false

for the STR_TO_DATE expression in the ORDER BY. This failure to equate
the two calls (and thus adding an extra item to fields) happens in
find_order_in_list. The failure happens because the order item hasn't been resolved
yet, making Item:eq fail to recognize the the two expressions are equal.

Solution: Resolve expressions before attempting find_item_in_list.
This also removes a redundant item for Query_block::fields.

Change-Id: I02e0bde33800ec5bfbbde7acce04494008152fb7