Bug#36483690 (noclose) "ALTER USER" fails in NDB schema distribution
This patch adds a new MTR test case stored_grants_set_pass,
which iteratively creates a random password on one server, then uses
it to authenticate to a different server.

The test runs in the ndb suite with ndb-log-bin disabled,
and also runs in the ndb_binlog suite with ndb-log-bin
and log-bin both enabled.

Change-Id: Ic7c4fc6ca026c31a433f79676522e979ec5beb49