Bug#36421351: Create new range for 9.x server-to-client error messages
Define a starting point for mysqld 9.x error messages to clients.
It is chosen so that the first 9.x error number is a fair bit ahead
of the last 8.x error number (ca. 250 at the time of this writing).
That way, we will have a "buffer" of unused 8.x error numbers that
we can use when we'll need to backport bugfixes that introduce new
messages from 9.x to 8.x.

The starting point was chosen by committee (Tatiana, Gopal, Praveen,
Sivert, Erlend) to account for the difficulty of choosing the ideal
number (providing a large enough reserve for backports, while not
wasting slots, seeing as error numbers below 10,000 are a somewhat
limited resource). This process should be easier for 10.x when
hopefully we'll have useful "backports / year" numbers from the
9.x -> 8.x versions.

Change-Id: I55276561e0dad237138460de925c7a420d2639ec