Bug#36487526 MDS Crash in X plugin protocol encoder
Also closes duplicate Bug#36505461.

When trying to return metadata, the code sees a visible column (the
first in the result set) which has no item_name set, which is
inconsistent; all visible items in the select list have item names
set.

The first column in the result set in this case (marked as visible) is
the COALESCE expression generated by the transform. This should have
been marked as a hidden item.

The underlying issue is that when the COALESCE is wrapped around the
expression in Item_singlerow_subselect::replace_scalar_subquery,
the new COALESCE doesn't inherit the hidden status of the expression
it is wrapping.

Solution: make COALESCE inherit the hidden status.
Change-Id: Ia44e605db378bfc46821af5cdfefdf90999033ab