Bug#36286802 RPM packaging does "make test" even when unit tests are disabled
Simplified patch for 8.0

Change the meaning of the EXCLUDE_FROM_PGO option for misc. cmake
macros/functions. We now skip building when generating profile data,
but we *do* build when *using* profile data. This is done in order to be
able to run unit tests also for PGO/LTO builds.

Change-Id: Ibb88d634ea02c14a4ae4c7372a11134195e8aec6