Bug#36364619: Assertion 'src != dst || cs->casedn_multiply == 1'
agg_item_set_converter() was recently updated to set the collation
for an argument item in some cases, however this can be dangerous
when the item's metadata relies upon collation data. This is e.g.
the case for the LOWER and UPPER functions, which set "multiply"
based on the collation. At runtime, the "multiply" argument is not
consistent with the actual collation, and the server gets in trouble.

The solution to this is to restrict setting collation to literal
strings only, since these are the only ones that really need it,
and does not have other metadata that depend upon the collation.

Also extend setting the collation when the desired collation is binary.
This was required due to failing range tests with the hypergraph
optimizer. It was also noticed that the collation field and
the collation of str_value inside class Item_string had to be
synchronized.

Finally, this caused problems for some statement digest tests when the
input string is binary. It was decided that this is not a reasonable
input type, so we now give an explicit error message for this case.

The original push caused one regression, which is fixed now.

Bug#36400256 revealed a problem with the NAME_CONST function.
This function does not have an own Item::Type enum, but would rather
impersonate as its wrapped literal, ie. with e.g. STRING_ITEM or
INT_ITEM. This might cause a bad down_cast in the new code. Luckily,
no code depended on the impersonation, so we could fix this by adding
a dedicated Item::Type enum NAME_CONST_ITEM for this function.
We could also simplify the resolver code for NAME_CONST due to this.
A regression test case has also been added.

Bug#36399759 is another apparent regression.
Because the patch may perform more rewrites to literal string values
in expressions, a difference in a generated column expression is
noticed in the output of SHOW CREATE TABLE and selection from
INFORMATION_SCHEMA.COLUMNS.
The regression is minor, since the rewite is now also performed when
the original character set is binary.
Investigations revealed that the problem stems from writing generation
expressions into the dictionary from unresolved items.
To silence the test case regression, the new result text is recorded
in this patch.

Change-Id: If4d062c5ff329f94b2c95078b4d1c595ca032cd5