Bug#36306144 escalation of Bug 35642087
Problem
===================

When a primary host experiences a network inactivity around 20 seconds,
all secondaries fail in a small chance.

Failure happens in
Gcs_xcom_communication_protocol_changer::decrement_nr_packets_in_transit()

Analysis and Fix
==================

We think that it might be related with a proposal made after it was
intended to be, thus warping the origin of the node, thus causing the failure.
We also think that this does not affect at all any data, because data
does not depend on the origin.  Only some internal algorithms that we
believe that is scarcely used:
- Dynamic Version Change and Group Replication AFTER consistency.

We created a two-fold patch that:
 - Drops the package if it detects a warped origin.
 - Prints debug information that helps to prove our theory.

Change-Id: I6ac29c661c17c060674f588d3abb79e32dbcdf66