WL#16179 Change default values for InnoDB variables [followup test fix]
Note: This commit is a followup test fix

Description:
Previously added option file sets innodb_flush_method to fsync. This
option is not supported on Windows and the server does not start.

Fix:
Removed option file as the O_DIRECT option is not supported on Windows.
Modified the test to expect O_DIRECT as a default along with fsync and
unbuffered.

Change-Id: Ibae8a663304d2de7be5a4537dd16dcb86cb37f09