Bug#36500214 MakeHypergraphTest.TraceBuffer may interact with other unit tests
The TraceBuffer class stores optimizer trace in segments of 4096 bytes.
When this trace is read, there is a check to decide if we have reached the
end of the trace. This check may give a false positive under the following
circumstances:

* The length of the trace is a multiple of 4096 bytes.
* The final segment is allocated in a region just before another segment.

Then std::streambuf::pptr() would point to last_segment.end() which happened
to be equal to other_segment.begin(). The test for reaching the end did not
handle this case correctly. This commit extends the at-the-end predicate
so that this error is fixed.

Change-Id: Ia2759a6f153c0abdc728ecf557d135dc67fe1d6e