Bug#36514834 NodeCertificate-t reports "Test 'lifetime' Failed: 11"
The method CertLifetime::set_set_cert_lifetime(X509 *cert) is
supposed to set the not-before and not-after times in cert to
the same as those stored in the CertLifetime object. But what it
actually did was to set not-before to the current time, and set
not-after to have the same duration as the object.

This showed up in test failures only if the real time clock ticked
a second while the test was running.

Fix the problem by reimplementing set_cert_lifetime() to use
ASN1_TIME_set() rather than X509_gmtime_adj().

Change-Id: If3539cc581f3d8124d8a999ea5a5d246be0d00d3