Fix crash in Item_cache::walk due to parameter comparison in nested subquery in stored proc
The pointer to the Item object in Item_cache object is never
cleared even after the Item object itself is rendered unusable
using the function TRASH. In cases where the same query is
run in the same MySQL thread but against a different database
the TRASHed Item object ends up being reused since its pointer
is valid and the system crashes because its data is not valid.

This commit fixes the bug https://bugs.mysql.com/bug.php?id=114657

This contribution is under the OCA signed by Amazon and covering
submissions to the MySQL project.