Bug# 36372334 - main.kill test fails after restoring MYSQL_OPT_RECONNECT
Problem:
After pushing WL#16221, the main.kill test had some issues related
to reconnect, mainly because of some reconnect issue in the mysqltest
client utility, so, the failing portion of the test was disabled to
avoid failures on pb2.

Fix:
Gerrit#29632 had pushed a fix for mysqltest client utility that fixes
the issue related to reconnect.

The test now passes without any failures.

Change-Id: Ife1d4dc3e6d1c67685c92949db250534c836cb1a